Cleaned up IRK.lbr (Eagle CAD device library)
Also, added a footprint for a very cheap micro-tactile switch
from Leye Electronics (China) - available on eBay. If you want
to use it you will have to redesign the PCB slightly as it is
a 6mm x 6mm SMT switch - larger than the C&K switch in the SMT
version.