Spaced the holes for the IR TX and Activity LEDs at 0.1 inch intervals.
Now (if you want) you can solder a 12x1 pin socket header and plug
your IR TX and Activity LEDs into it rather than soldering them
directly to the board. Long term reliability might be an issue
as the leads oxidise though. The interspersed pin sockets appear on
the schematic as grounded 1x1 header pins.

The IR RX LED has remained slightly offset from the others, because it
needs to be bent 90° so that it is perpendicular to the board.

The board version is now 2.01 as a result of these minor layout changes
and the Gerber files have been regenerated.