Updated the LCD package to have 3 mm mounting holes.
The original ones were 2.5 mm but didn't match reality.
So I had to drill out the mounting holes to 3 mm (or else
find 2.5 mm bolts which are quite hard to find). M3 bolts
are much more commonly available. Had to reroute one of the
PCB traces to avoid a hole as a result.